feat: add trail that follows user clicks in MoodEnergyChart
* Upload MoodEnergyChart code

Uploading MoodEnergyChart component and test files, versions that implement the Trail and the history of the chart and earlier moods.

* Minor updates to both pages

There is only one line of code changed in each page, and that is in the use of the MoodEnergyChart. LandingPage has "trailEnabled" set to False, while ExplorePage as "trailEnabled" set to True. This will make it so the trail/history appears in ExplorePage, but not in LandingPage.

* fix: import Line from recharts

* fix: restore previously deleted tests

---------

Co-authored-by: Danny Semakov <semakovdaniel@gmail.com>