feat: make song recommendations based on mood and genre significantly more accurate using new song data set
Switched from the ReccoBeats API, which gave song recommendations in a very unintuitive way, to a custom-made API using a data set of 30,000 songs. Removed code mentioning ReccoBeats API. Allowed user to select a genre from the dropdown menu on the landing page and recommended songs within that genre. Some recommended songs do not intuitively fit within the selected genre, but this seems to be a problem with the data set rather than the code.