feat: allow user to get spotify playlist recommendations (rather than individual songs) and significantly improve the format for the explore page
Created an endpoint that takes in valence, energy, and mood then creates a matching spotify playlist and returns info about the playlist. Created a frontend function to call this endpoint and then pass this playlist to the TrackDisplay component to embed the playlist. Removed arrow buttons and horizontal scroll from explore page. Genre dropdown, chart, and track display are all displayed side-by-side and fit on the page for all common screen sizes.